#913: ToDoBatchItem 공통 엔티티로 분리
- TDFoundation에 SharedEntity를 추가해 공통 엔티티인 ToDoBatchItem을 분리하였습니다.
